AJUESTES 30082021
AJUSTES 30082021

+++ application/controllers/insumos/cprog_insumo.php
@@ -189,7 +189,7 @@ class Cprog_insumo extends CI_Controller{
                     $unidad = utf8_encode(trim($datos[3])); //// Unidad
                     $cantidad = intval(trim($datos[4])); //// Cantidad
                     $unitario = intval(trim($datos[5])); //// Costo Unitario
-                    $total = intval(trim($datos[6])); //// Costo Total
+                    $total =  floatval(trim($datos[6])); //// Costo Total
 
                     $p_total=round(($cantidad*$unitario),2);
